correct BinaryFile max length and optimize
Buffer was read with 100000, no need to expand to 1000000,
and if 100000 is enough for reading a file it should be enough
for any image passed in through a buffer, otherwise it needs to
be increased in both places.  Adding a variable to use the same size
in all three locations.

Allocate a buffer no bigger than the current buffer, rather than
always the full size.